cargo_auto_template_new_pwa_wasm::main_mod
Module lib_mod

This module is like a lib.rs module for a binary CLI executable. The lib_mod.rs must not contains any input/output interface stuff. So the program logic can be separate from the interface.

Re-exports

pub use hello_mod::format_hello_phrase;

pub use hello_mod::format_upper_hello_phrase;

pub use web_sys_mod as wsm;

Modules

hello_mod 🔒

All the real code (program logic) is inside modules in separate files. This module are UI agnostic and must not have anything to do with UI. So the same library could be used for CLI and for WASM, that have vastly different UI.

web_sys_mod

Helper functions for web_sys, window, document, dom, console, html elements,…
Trying to isolate/hide all javascript code and conversion in this module.

Enums

LibraryError

all possible library errors for thiserror

Constants

GREEN

RED

RESET

YELLOW