cargo_auto_template_new_pwa_wasm::main_mod::lib_mod::web_sys_mod
Module html_source_code_mod

html_source_code_mod.rs

Structs

HtmlSourceCode

HtmlSourceCode - type to manipulate HTML source code safer than with string functions only
WARNING for HTML INJECTION!
HTML is the standard markup language for Web pages. HTML source code is just a text.
It is easy to read, write, understand and parse.
The syntax of HTML source code is similar to XML structured with elements, tags, nodes, texts, comments and attributes.
The browser then transforms this HTML source code into the DOM tree and then renders that.
It is very tempting to modify this source code in our application with string manipulation and then pass it to the browser.
The html source code (it is just a string) that is provided by the programmer is always ok, he wants it to work properly.
The BIG problem arises when we need to inject some user provided data into the HTML source code.
The HTML syntax mixes instructions and data together and this creates a BIG problem.
Never put user provided strings in a html source code directly, because it can contain an HTML injection attack.
We need to encode all user data before putting it into the HTML source code.
There are 2 types of encodings: one for attributes values and another for text nodes.
We will create a new type that makes it safer and easier for the programmer to replace data in the HTML source code.